server/requirements.txt
aiocache==0.10.1
aiohttp==3.5.2
aiohttp-cors==0.7.0
aiopg==0.16.0
aioredis==1.2.0
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
future==0.17.1
hiredis==1.0.0
idna==2.8
jsonschema==3.0.1
msgpack==0.6.1
multidict==4.5.2
numpy==1.16.1
pandas==0.24.1
psycopg2-binary==2.7.7
PyJWT==1.7.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
six==1.12.0
SQLAlchemy==1.2.18
ujson==1.35
wcwidth==0.1.7
XlsxWriter==1.1.5
yarl==1.3.0
